requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
pymongo==4.4.0
python-dotenv==1.0.0
redis==5.0.0
neo4j==5.28.1

